Update logic to extract details from git_config
Extract the user & repo (& project for codebasehq) based on the
type of git repository we have as the format in git_config is different
if we're looking at an SSH config or an HTTP one.

The format for non-codebasehq repositories is:
SSH: {example.com}:{user}/{repo}.git
HTTP: {example.com}/{user}/{repo}.git

The codebasehq format is slightly different as we need the project and
the user name is in a different place.
SSH: "codebasehq.com:{user}/{project}/{repo}.git"
HTTP: "{user}.codebasehq.com/{project}/{repo}.git"

Also swap the test for codebasehq to a positive check for clarity.